Reuse are_equal_ignoring_line_endings in another test
